[utils] remove hardware dispatch from attention padding helpers
Move the existing device-agnostic PyTorch/einops implementation into attention_utils and remove the NPU-only duplicate.
