+++ verl/utils/attention_utils.py
@@ -12,89 +12,84 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Callable
-
-_index_first_axis, _pad_input, _rearrange, _unpad_input = None, None, None, None
-
-
-def _get_attention_functions() -> tuple[Callable, Callable, Callable, Callable]:
-    """Dynamically import attention functions based on available hardware."""
-
-    from verl.utils.device import is_torch_npu_available
-
-    global _index_first_axis, _pad_input, _rearrange, _unpad_input
-
-    if is_torch_npu_available(check_device=False):
-        from verl.utils.npu_flash_attn_utils import index_first_axis, pad_input, rearrange, unpad_input
-    else:
-        from flash_attn.bert_padding import index_first_axis, pad_input, rearrange, unpad_input
-
-    _index_first_axis, _pad_input, _rearrange, _unpad_input = index_first_axis, pad_input, rearrange, unpad_input
-
-    return _index_first_axis, _pad_input, _rearrange, _unpad_input
+import torch
+import torch.nn.functional as F
+from einops import rearrange, repeat
 
 
-def index_first_axis(*args, **kwargs):
-    """
-    Unified entry point for `index_first_axis` across CUDA and NPU backends.
+# Adapted from https://github.com/Dao-AILab/flash-attention/blob/main/flash_attn/bert_padding.py.
+# These operations use only PyTorch and einops, so the same implementation works on every PyTorch device.
+class IndexFirstAxis(torch.autograd.Function):
+    @staticmethod
+    def forward(ctx, input, indices):
+        ctx.save_for_backward(indices)
+        assert input.ndim >= 2
+        ctx.first_axis_dim, other_shape = input.shape[0], input.shape[1:]
+        second_dim = other_shape.numel()
+        return torch.gather(
+            rearrange(input, "b ... -> b (...)"),
+            0,
+            repeat(indices, "z -> z d", d=second_dim),
+        ).reshape(-1, *other_shape)
 
-    Dynamically dispatches to the appropriate device-specific implementation:
-      - On CUDA: `flash_attn.bert_padding.index_first_axis`
-      - On NPU: `transformers.integrations.npu_flash_attention.index_first_axis`
-        (falls back to `transformers.modeling_flash_attention_utils._index_first_axis`
-        in newer versions of transformers).
-
-    Users can call this function directly without worrying about the underlying device.
-    """
-    func, *_ = _get_attention_functions()
-    return func(*args, **kwargs)
-
-
-def pad_input(*args, **kwargs):
-    """
-    Unified entry point for `pad_input` across CUDA and NPU backends.
-
-    Dynamically dispatches to the appropriate device-specific implementation:
-      - On CUDA: `flash_attn.bert_padding.pad_input`
-      - On NPU: `transformers.integrations.npu_flash_attention.pad_input`
-        (falls back to `transformers.modeling_flash_attention_utils._pad_input`
-        in newer versions of transformers).
-
-    Users can call this function directly without worrying about the underlying device.
-    """
-    _, func, *_ = _get_attention_functions()
-    return func(*args, **kwargs)
+    @staticmethod
+    def backward(ctx, grad_output):
+        (indices,) = ctx.saved_tensors
+        assert grad_output.ndim >= 2
+        other_shape = grad_output.shape[1:]
+        grad_output = rearrange(grad_output, "b ... -> b (...)")
+        grad_input = torch.zeros(
+            [ctx.first_axis_dim, grad_output.shape[1]],
+            device=grad_output.device,
+            dtype=grad_output.dtype,
+        )
+        grad_input.scatter_(0, repeat(indices, "z -> z d", d=grad_output.shape[1]), grad_output)
+        return grad_input.reshape(ctx.first_axis_dim, *other_shape), None
 
 
-def rearrange(*args, **kwargs):
-    """
-    Unified entry point for `rearrange` across CUDA and NPU backends.
-
-    Dynamically dispatches to the appropriate device-specific implementation:
-      - On CUDA: `flash_attn.bert_padding.rearrange`
-      - On NPU: `transformers.integrations.npu_flash_attention.rearrange`
-        (falls back to `einops.rearrange` if no dedicated NPU implementation exists).
-
-    Users can call this function directly without worrying about the underlying device.
-    """
-    *_, func, _ = _get_attention_functions()
-    return func(*args, **kwargs)
+index_first_axis = IndexFirstAxis.apply
 
 
-def unpad_input(*args, **kwargs):
-    """
-    Unified entry point for `unpad_input` across CUDA and NPU backends.
+class IndexPutFirstAxis(torch.autograd.Function):
+    @staticmethod
+    def forward(ctx, values, indices, first_axis_dim):
+        ctx.save_for_backward(indices)
+        assert indices.ndim == 1
+        assert values.ndim >= 2
+        output = torch.zeros(first_axis_dim, *values.shape[1:], device=values.device, dtype=values.dtype)
+        output[indices] = values
+        return output
 
-    Dynamically dispatches to the appropriate device-specific implementation:
-      - On CUDA: `flash_attn.bert_padding.unpad_input`
-      - On NPU: `transformers.integrations.npu_flash_attention.unpad_input`
-        (falls back to `transformers.modeling_flash_attention_utils._unpad_input`
-        in newer versions of transformers).
+    @staticmethod
+    def backward(ctx, grad_output):
+        (indices,) = ctx.saved_tensors
+        return grad_output[indices], None, None
 
-    Users can call this function directly without worrying about the underlying device.
-    """
-    *_, func = _get_attention_functions()
-    return func(*args, **kwargs)
+
+index_put_first_axis = IndexPutFirstAxis.apply
+
+
+def pad_input(hidden_states, indices, batch, seqlen):
+    """Restore unpadded hidden states to shape ``(batch, seqlen, ...)``."""
+    output = index_put_first_axis(hidden_states, indices, batch * seqlen)
+    return rearrange(output, "(b s) ... -> b s ...", b=batch)
+
+
+def unpad_input(hidden_states, attention_mask, unused_mask=None):
+    """Remove padding and return FlashAttention-compatible sequence metadata."""
+    all_masks = (attention_mask + unused_mask) if unused_mask is not None else attention_mask
+    seqlens_in_batch = all_masks.sum(dim=-1, dtype=torch.int32)
+    used_seqlens_in_batch = attention_mask.sum(dim=-1, dtype=torch.int32)
+    indices = torch.nonzero(all_masks.flatten(), as_tuple=False).flatten()
+    max_seqlen_in_batch = seqlens_in_batch.max().item()
+    cu_seqlens = F.pad(torch.cumsum(seqlens_in_batch, dim=0, dtype=torch.int32), (1, 0))
+    return (
+        index_first_axis(rearrange(hidden_states, "b s ... -> (b s) ..."), indices),
+        indices,
+        cu_seqlens,
+        max_seqlen_in_batch,
+        used_seqlens_in_batch,
+    )
 
 
 __all__ = ["index_first_axis", "pad_input", "rearrange", "unpad_input"]
